Restructure files to include folders for redux and pointing to entry_level file
