--- client/MobileApp/src/data/reducer.js
@@ -1,6 +1,0 @@
-import { combineReducers } from 'redux';
-import { reducer as usersReducer } from './users/reducer';
-
-export const reducer = combineReducers({
-	users: usersReducer,
-});
