--- client/MobileApp/src/data/users/actionTypes.js
@@ -1,2 +1,0 @@
-export const UPDATE = 'users/UPDATE';
-export const EMPTY = 'users/EMPTY';
