--- client/MobileApp/src/data/users/api.js
@@ -1,10 +1,0 @@
-import { fetchApi } from 'MobileApp/src/services/api';
-
-const endPoints = {
-	create: '/users',
-	get: '/users',
-};
-
-export const create = payload => fetchApi(endPoints.create, payload, 'post');
-
-export const get = payload => fetchApi(endPoints.get, payload, 'get');
