--- client/MobileApp/src/data/users/reducer.js
@@ -1,26 +1,0 @@
-import * as actionTypes from './actionTypes';
-
-const initialState = {
-	items: {},
-};
-
-export const reducer = (state = initialState, action) => {
-	switch (action.type) {
-		case actionTypes.UPDATE:
-			return {
-				items: {
-					...state.items,
-					...action.items.reduce((prev, curr) => ({
-						...prev,
-						[curr.id]: curr,
-					}), {}),
-				},
-			};
-		case actionTypes.EMPTY:
-			return {
-				items: {},
-			};
-		default:
-			return state;
-	}
-};
