--- client/MobileApp/src/services/api/config.js
@@ -1,4 +1,0 @@
-export default {
-	clientId: '8puWuJWZYls1Ylawxm6CMiYREhsGGSyw',
-  url: 'http://10.163.109.163:1337',
-};
