--- client/MobileApp/src/services/persist/actionTypes.js
@@ -1,1 +1,0 @@
-export const UPDATE = 'persist/UPDATE';
