--- client/MobileApp/src/services/persist/actions.js
@@ -1,6 +1,0 @@
-import * as actionTypes from './actionTypes';
-
-export const update = payload => ({
-	type: actionTypes.UPDATE,
-	payload,
-});
