--- client/MobileApp/src/services/persist/reducer.js
@@ -1,15 +1,0 @@
-import * as actionTypes from './actionTypes';
-
-export const initialState = {
-	isHydrated: false,
-};
-
-
-export function reducer(state = initialState, action) {
-	switch (action.type) {
-		case actionTypes.UPDATE:
-			return action.payload;
-		default:
-			return state;
-	}
-}
