--- client/MobileApp/src/services/routeHistory/actionTypes.js
@@ -1,1 +1,0 @@
-export const PUSH = 'routeHistory/PUSH';
