--- client/MobileApp/src/services/routeHistory/actions.js
@@ -1,6 +1,0 @@
-import * as actionTypes from './actionTypes';
-
-export const push = route => ({
-	type: actionTypes.PUSH,
-	route,
-});
