--- client/MobileApp/src/services/session/actionTypes.js
@@ -1,1 +1,0 @@
-export const UPDATE = 'session/UPDATE';
