--- client/MobileApp/src/services/session/actions.js
@@ -1,6 +1,0 @@
-import * as actionTypes from './actionTypes';
-
-export const update = session => ({
-	type: actionTypes.UPDATE,
-	session,
-});
