--- client/MobileApp/src/services/session/selectors.js
@@ -1,3 +1,0 @@
-import store from 'MobileApp/src/store';
-
-export const get = () => store.getState().services.session;
