+++ client/MobileApp/store/store.js
@@ -1,0 +1,14 @@
+// import { createStore, applyMiddleware } from 'redux';
+// import thunk from 'redux-thunk';
+
+// import rootReducer from '../reducers/root_reducer';
+
+// const configureStore = (preloadedState = {}) => (
+//   createStore(
+//     rootReducer, 
+//     preloadedState, 
+//     applyMiddleware(thunk)
+//   )
+// );
+
+// export default configureStore;
