+++ client/MobileApp/util/api_util.js
@@ -1,0 +1,70 @@
+import { AsyncStorage } from 'react-native';
+
+
+export const createUser = (username, password) => (
+  fetch("https://kitchenfox.herokuapp.com/api/register", {
+    method: "POST",
+    headers: {
+      'Content-Type': 'application/x-www-form-urlencoded',
+    },
+    body: `username=${username}&password=${password}`,
+  })
+);
+
+export const login = (username, password) => (
+  fetch("https://kitchenfox.herokuapp.com/api/login", {
+    method: "POST",
+    headers: {
+      'Content-Type': 'application/x-www-form-urlencoded',
+      charset: 'UTF-8',
+    },
+    body: `username=${username}&password=${password}`,
+  })
+);
+
+export const saveToken = (response) => {
+  let body = JSON.parse(response._bodyText);
+  let token = body.token;
+  console.log(token);
+  AsyncStorage.setItem('jwt', JSON.stringify(token));
+};
+
+export const getLocalToken = () => (
+  AsyncStorage.getItem('jwt')
+);
+
+export const demoSecured = () => {
+  const token = getLocalToken();
+  return (
+    fetch('https://kitchenfox.herokuapp.com/api/protected', {
+      method: 'GET',
+      headers: {
+        authorization: `JWT ${token}`,
+        'Content-Type': 'application/x-www-form-urlencoded',
+        charset: 'UTF-8',
+      },
+    })
+  );
+}
+
+export const securable = token => (
+  fetch("https://kitchenfox.herokuapp.com/api/protected", {
+    method: "GET",
+    headers: {
+      authorization: `JWT ${token}`,
+      'Content-Type': 'application/x-www-form-urlencoded',
+      charset: 'UTF-8',
+    },
+  })
+);
+
+export const protectedHeaders = () => (
+  fetch("https://kitchenfox.herokuapp.com/api/protected", {
+    method: "GET",
+    headers: {
+      authorization: `JWT ${token}`,
+      'Content-Type': 'application/x-www-form-urlencoded',
+      charset: 'UTF-8',
+    },
+  })
+);
